refactor: dedupe test_all_to_all helpers into shared module (-557/+245 lines across 2 copies)
Move run_with_trace/run_all_to_all_impl from box/all_post_commit and
box/nightly into blackhole_CI/_all_to_all_helpers.py. The copies were
identical except for the CacheEntriesCounter import location (module level vs
inside the impl) — no behavior parameters needed. A dead first run_with_trace
definition, immediately shadowed by a second one in both copies, is dropped.
Tails unchanged.
