+++ tests/ttnn/unit_tests/operations/ccl/blackhole_CI/box/all_post_commit/test_all_to_all.py
@@ -4,288 +4,10 @@
 
 import pytest
 import ttnn
-import torch
-from loguru import logger
 
-from tests.tt_eager.python_api_testing.sweep_tests.comparison_funcs import comp_equal, comp_pcc
 from tests.ttnn.unit_tests.operations.ccl.blackhole_CI.box.nightly.test_all_gather_nightly import validate_test
 from models.common.utility_functions import skip_for_wormhole_b0, skip_for_n_or_less_dev
-
-
-def run_with_trace(
-    mesh_device,
-    all_to_all_topology,
-    input_tensor_mesh,
-    in_dim,
-    out_dim,
-    num_links,
-    output_mem_config,
-    multi_device_global_semaphore,
-    num_iter=20,
-    subdevice_id=None,
-):
-    # Compile Run
-    logger.info("Compiling model")
-    tt_out_tensor = ttnn.experimental.all_to_all_async(
-        input_tensor_mesh,
-        in_dim,
-        out_dim,
-        multi_device_global_semaphore=multi_device_global_semaphore,
-        num_links=num_links,
-        memory_config=output_mem_config,
-        topology=all_to_all_topology,
-        subdevice_id=subdevice_id,
-    )
-    ttnn.synchronize_device(mesh_device)
-
-    # Capture trace
-    logger.info("Capturing trace")
-    output_tensors = []
-    trace_id = ttnn.begin_trace_capture(mesh_device, cq_id=0)
-    for i in range(num_iter):
-        tt_out_tensor = ttnn.experimental.all_to_all_async(
-            input_tensor_mesh,
-            in_dim,
-            out_dim,
-            multi_device_global_semaphore=multi_device_global_semaphore,
-            num_links=num_links,
-            memory_config=output_mem_config,
-            topology=all_to_all_topology,
-            subdevice_id=subdevice_id,
-        )
-        output_tensors.append(tt_out_tensor)
-    ttnn.end_trace_capture(mesh_device, trace_id, cq_id=0)
-    ttnn.synchronize_device(mesh_device)
-
-    # Run the op
-    logger.info("Starting Trace perf test...")
-    ttnn.execute_trace(mesh_device, trace_id, blocking=False)
-    ttnn.release_trace(mesh_device, trace_id)
-    ttnn.synchronize_device(mesh_device)
-
-    return output_tensors
-
-
-def run_with_trace(
-    mesh_device,
-    topology,
-    input_tensor,
-    persistent_intermediate_buffer,
-    persistent_output_buffer,
-    in_dim,
-    out_dim,
-    num_links,
-    output_mem_config,
-    multi_device_global_semaphore,
-    num_iter=20,
-    subdevice_id=None,
-):
-    # Compile Run
-    logger.info("Compiling model")
-    tt_out_tensor = ttnn.experimental.all_to_all_async(
-        input_tensor,
-        in_dim=in_dim,
-        out_dim=out_dim,
-        persistent_intermediate_buffer=persistent_intermediate_buffer,
-        persistent_output_buffer=persistent_output_buffer,
-        multi_device_global_semaphore=multi_device_global_semaphore,
-        num_links=num_links,
-        memory_config=output_mem_config,
-        topology=topology,
-        subdevice_id=subdevice_id,
-    )
-    ttnn.synchronize_device(mesh_device)
-
-    # Capture trace
-    logger.info("Capturing trace")
-    trace_id = ttnn.begin_trace_capture(mesh_device, cq_id=0)
-    for i in range(num_iter):
-        tt_out_tensor = ttnn.experimental.all_to_all_async(
-            input_tensor,
-            in_dim=in_dim,
-            out_dim=out_dim,
-            persistent_intermediate_buffer=persistent_intermediate_buffer,
-            persistent_output_buffer=persistent_output_buffer,
-            multi_device_global_semaphore=multi_device_global_semaphore,
-            num_links=num_links,
-            memory_config=output_mem_config,
-            topology=topology,
-            subdevice_id=subdevice_id,
-        )
-    ttnn.end_trace_capture(mesh_device, trace_id, cq_id=0)
-    ttnn.synchronize_device(mesh_device)
-
-    # Run the op
-    logger.info("Starting Trace perf test...")
-    ttnn.execute_trace(mesh_device, trace_id, blocking=False)
-    ttnn.release_trace(mesh_device, trace_id)
-    ttnn.synchronize_device(mesh_device)
-
-    return tt_out_tensor
-
-
-def run_all_to_all_impl(
-    mesh_device,
-    num_devices,
-    logical_shape,
-    in_dim,
-    out_dim,
-    num_links,
-    input_dtype,
-    layout,
-    topology,
-    num_iters=1,
-    trace_mode=False,
-    mem_config=None,
-    do_check=True,
-    reuse_inputs=False,
-):
-    if num_iters < 1:
-        pytest.fail("num_iters must be >= 1")
-
-    from tests.tests_common.cache_entries_counter import CacheEntriesCounter
-
-    mesh_device.cache_entries_counter = CacheEntriesCounter(mesh_device)
-
-    compute_grid_size = mesh_device.compute_with_storage_grid_size()
-    ccl_sub_device_crs = ttnn.CoreRangeSet(
-        {ttnn.CoreRange(ttnn.CoreCoord(0, 0), ttnn.CoreCoord(compute_grid_size.x - 1, compute_grid_size.y - 1))}
-    )
-    worker_sub_device = ttnn.SubDevice(
-        [
-            ccl_sub_device_crs,
-        ]
-    )
-    worker_sub_device_id = ttnn.SubDeviceId(0)
-    sub_device_stall_group = [worker_sub_device_id]
-    sub_device_manager = mesh_device.create_sub_device_manager([worker_sub_device], 0)
-    mesh_device.load_sub_device_manager(sub_device_manager)
-    mesh_device.set_sub_device_stall_group(sub_device_stall_group)
-    # create global semaphore handles
-    ccl_semaphore_handles = [ttnn.create_global_semaphore(mesh_device, ccl_sub_device_crs, 0) for _ in range(num_iters)]
-
-    logger.info(f"Logical shape: {logical_shape}")
-    logger.info(f"in_dim: {in_dim}")
-    logger.info(f"out_dim: {out_dim}")
-
-    input_mem_config = mem_config
-    output_mem_config = mem_config
-    ###
-
-    ### Create persistent output buffers
-    logger.info("Creating persistent buffers")
-    output_shape = list(logical_shape)
-    output_shape[out_dim] //= num_devices
-    persistent_intermediate_buffers = [
-        ttnn.from_torch(
-            torch.zeros(output_shape),
-            device=mesh_device,
-            layout=ttnn.TILE_LAYOUT,
-            dtype=input_dtype,
-            memory_config=output_mem_config,
-            mesh_mapper=ttnn.ReplicateTensorToMesh(mesh_device),
-        )
-        for _ in range(num_iters)
-    ]
-    persistent_output_buffers = [
-        ttnn.from_torch(
-            torch.zeros(output_shape),
-            device=mesh_device,
-            layout=ttnn.TILE_LAYOUT,
-            dtype=input_dtype,
-            memory_config=output_mem_config,
-            mesh_mapper=ttnn.ReplicateTensorToMesh(mesh_device),
-        )
-        for _ in range(num_iters)
-    ]
-
-    logger.info("Done creating persistent buffers")
-
-    input_tensor_mesh_list = []
-    output_tensor_goldens_list = []
-
-    for i in range(num_iters if not reuse_inputs else 1):
-        output_tensor = torch.rand(logical_shape).bfloat16()
-        output_tensor_goldens_list.append(torch.chunk(output_tensor, num_devices, out_dim))
-        input_tensor_mesh = ttnn.from_torch(
-            output_tensor,
-            device=mesh_device,
-            layout=layout,
-            dtype=input_dtype,
-            memory_config=input_mem_config,
-            mesh_mapper=ttnn.create_mesh_mapper(
-                mesh_device,
-                ttnn.MeshMapperConfig(
-                    [ttnn.PlacementReplicate(), ttnn.PlacementShard(in_dim)], ttnn.MeshShape(1, num_devices)
-                ),
-            ),
-        )
-
-        input_tensor_mesh_list.append(input_tensor_mesh)
-
-    tt_out_tensor_list = []
-    with mesh_device.cache_entries_counter.measure():
-        if trace_mode:
-            tt_out_tensor = run_with_trace(
-                mesh_device,
-                topology,
-                input_tensor_mesh_list[0],
-                persistent_intermediate_buffers[0],
-                persistent_output_buffers[0],
-                in_dim,
-                out_dim,
-                num_links,
-                output_mem_config,
-                multi_device_global_semaphore=ccl_semaphore_handles[0],
-                num_iter=num_iters,
-                subdevice_id=worker_sub_device_id,
-            )
-            tt_out_tensor_list.append(tt_out_tensor)
-        else:
-            for i in range(num_iters):
-                tt_out_tensor = ttnn.experimental.all_to_all_async(
-                    input_tensor_mesh_list[i if not reuse_inputs else 0],
-                    persistent_intermediate_buffer=persistent_intermediate_buffers[i],
-                    persistent_output_buffer=persistent_output_buffers[i],
-                    in_dim=in_dim,
-                    out_dim=out_dim,
-                    multi_device_global_semaphore=ccl_semaphore_handles[i],
-                    num_links=num_links,
-                    memory_config=output_mem_config,
-                    topology=topology,
-                    subdevice_id=worker_sub_device_id,
-                )
-                tt_out_tensor_list.append(persistent_output_buffers[i])
-
-            logger.info(f"Waiting for op")
-            ttnn.synchronize_device(mesh_device, sub_device_ids=sub_device_stall_group)
-            logger.info(f"Done op")
-
-    passed = True
-    if do_check:
-        for tensor_index in range(len(tt_out_tensor_list)):
-            tt_out_tensor = tt_out_tensor_list[tensor_index]
-            output_tensors = output_tensor_goldens_list[tensor_index if not reuse_inputs else 0]
-            for i, t in enumerate(ttnn.get_device_tensors(tt_out_tensor)):
-                tt_output_tensor = t.cpu().to(ttnn.ROW_MAJOR_LAYOUT).to_torch()
-                output_tensor = output_tensors[i]
-                logger.info(f"Checking for device {t.device().id()}")
-                if input_dtype == ttnn.bfloat16:
-                    eq, output = comp_equal(tt_output_tensor, output_tensor)
-                else:
-                    eq, output = comp_pcc(tt_output_tensor, output_tensor)
-                if not eq:
-                    logger.error(f"output mismatch for tensor {i}")
-                    passed = False
-
-    assert (
-        mesh_device.cache_entries_counter.total == 1
-    ), f"Device has {mesh_device.cache_entries_counter.total} program cache entries"
-
-    mesh_device.reset_sub_device_stall_group()
-    mesh_device.clear_loaded_sub_device_manager()
-    if do_check and not passed:
-        assert eq, f"{i} FAILED: {output}"
+from tests.ttnn.unit_tests.operations.ccl.blackhole_CI._all_to_all_helpers import run_all_to_all_impl
 
 
 @skip_for_wormhole_b0()
